// SDKs/twilio.com/twilio_autopilot_v1/1.37.3/typescript/src/sdk/models/shared/autopilotv1assistantmodelbuild.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { ModelBuildEnumStatusEnum } from "./modelbuildenumstatusenum";



export class AutopilotV1AssistantModelBuild extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=assistant_sid" })
  assistantSid?: string;

  @SpeakeasyMetadata({ data: "json, name=build_duration" })
  buildDuration?: number;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=error_code" })
  errorCode?: number;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=status" })
  status?: ModelBuildEnumStatusEnum;

  @SpeakeasyMetadata({ data: "json, name=unique_name" })
  uniqueName?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}