// SDKs/twilio.com/twilio_autopilot_v1/1.37.3/typescript/src/sdk/models/operations/listmodelbuild.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import * as shared from "../shared";

export const ListModelBuildServerList = [
	"https://autopilot.twilio.com",
] as const;


export class ListModelBuildPathParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=AssistantSid" })
  assistantSid: string;
}


export class ListModelBuildQueryParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListModelBuildSecurity extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListModelBuildListModelBuildResponseMeta extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=first_page_url" })
  firstPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=key" })
  key?: string;

  @SpeakeasyMetadata({ data: "json, name=next_page_url" })
  nextPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=page_size" })
  pageSize?: number;

  @SpeakeasyMetadata({ data: "json, name=previous_page_url" })
  previousPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}


export class ListModelBuildListModelBuildResponse extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=meta" })
  meta?: ListModelBuildListModelBuildResponseMeta;

  @SpeakeasyMetadata({ data: "json, name=model_builds", elemType: shared.AutopilotV1AssistantModelBuild })
  modelBuilds?: shared.AutopilotV1AssistantModelBuild[];
}


export class ListModelBuildRequest extends SpeakeasyBase {
  @SpeakeasyMetadata()
  serverUrl?: string;

  @SpeakeasyMetadata()
  pathParams: ListModelBuildPathParams;

  @SpeakeasyMetadata()
  queryParams: ListModelBuildQueryParams;

  @SpeakeasyMetadata()
  security: ListModelBuildSecurity;
}


export class ListModelBuildResponse extends SpeakeasyBase {
  @SpeakeasyMetadata()
  contentType: string;

  @SpeakeasyMetadata()
  listModelBuildResponse?: ListModelBuildListModelBuildResponse;

  @SpeakeasyMetadata()
  statusCode: number;
}
